+++ CHANGELOG.md
@@ -1,3 +1,29 @@
+## 3.0.1
+
+Changes:
+- Set flutter constraints to ^3.22.0
+
+## 3.0.0
+
+Changes:
+- Change sdk constraints to ^3.3.0
+
+## 2.0.1
+
+Fixes:
+- Fix last tab selection
+
+## 2.0.0
+
+Features:
+- Add 'tabAlignment' parameter
+
+Fixes:
+- Fix last tab selection
+
+Changes:
+- Change sdk constraints to ">=3.2.0 <4.0.0"
+
 ## 1.1.2
 
 Features: